+++ spu/psi_pb2.py
@@ -1,1 +1,0 @@
-from libspu.psi.psi_pb2 import *